Fix and revamp final image generation.
Instead of generating multiple separate images, use a single image into
which all files are copied, and delegate launching them correctly to the
run target in the Makefile (which is why the entrypoint for the image is
now make, using JSON syntax per [1]). The only run target that isn't
ready for this is the SPIxCONV one.

Using a generic is a big change in behavior for the mbtemp IOC, which
attempted to update the IOC definitions using the spreadsheet before
actually launching them. This use case is no longer supported,
especially considering that it encourages haphazard untracked updates of
the spreadsheet.

This change also removes the image push makefile targets, since their
versions are specified in an .env file which isn't even sourced by the
Makefile, and we can use 'docker compose push' commands instead.

We would have liked to use a single 'docker compose build' command, but
the necessary functionality for expressing dependencies [2] is broken
for some versions of docker compose still in use [3].

[1] https://docs.docker.com/reference/build-checks/json-args-recommended/
[2] https://github.com/docker/compose/issues/12579
[3] https://docs.docker.com/compose/how-tos/dependent-images/