Update files generated with 'make db'.
The command was run with the spreadsheet checked into the repository, so
new files (beyond mbtemp cmd files, which are currently expected to be
generated at runtime, so weren't included in this commit) or big changes
weren't expected nor found.

The main change here is in the countingpru IOCs, where the Python script
changed PV prefixes from 'None' to 'nan'. The PV name change in one of
the mks937b IOCs was already applied in the currently running version of
the IOC.